Got the basic Tufte CSS tools incorporated.
I feel like the Tufte CSS tools are having a bit of a fight with the Jekyll "default" SASS/CSS files (especially on things like page width), so I need to do some additionally fiddling to figure that out.

Also, I'm not really using the Tufte tools in things like the syllabus to their full extent.